Move lookahead into own file
